suppressing error output and adding dnsmasq conf

+++ kong/cli/utils/signal.lua
@@ -228,7 +228,9 @@ function _M.send_signal(args_config, signal)
       cutils.logger:warn("Can't find dnsmasq. Using default resolver address")
     else
       -- Kill it first
-      IO.os_execute("kill `cat "..constants.CLI.DNSMASQ_PID.."`")
+      if IO.file_exists(constants.CLI.DNSMASQ_PID) then
+        IO.os_execute("kill `cat "..constants.CLI.DNSMASQ_PID.."`")
+      end
       local res, code = IO.os_execute(cmd.." -p "..kong_config.dnsmasq_port.." --pid-file="..constants.CLI.DNSMASQ_PID)
       if code ~= 0 then
         cutils.logger:error_exit(res)
@@ -240,9 +242,11 @@ function _M.send_signal(args_config, signal)
 
   if signal == STOP then
     -- Terminate dnsmasq
-    local _, code = IO.os_execute("kill `cat "..constants.CLI.DNSMASQ_PID.."`")
-    if code == 0 then
-      cutils.logger:info("dnsmasq stopped")
+    if IO.file_exists(constants.CLI.DNSMASQ_PID) then
+      local _, code = IO.os_execute("kill `cat "..constants.CLI.DNSMASQ_PID.."`")
+      if code == 0 then
+        cutils.logger:info("dnsmasq stopped")
+      end
     end
   end
